Refactor out configuration management
