+++ trieste.py
@@ -26,21 +26,24 @@
 
 from __future__ import division, print_function, unicode_literals
 
-import json
 import logging.handlers
+from collections import namedtuple
+from batch_shipyard import convoy
 
 import click
-import jsonschema
-from types import FunctionType as Func
-from pymonad.Maybe import *
-from collections import namedtuple
 
 try:
     import pathlib
 except ImportError:
     import pathlib2 as pathlib
 
-from batch_shipyard.convoy import *
+
+import configurations
+
+import sys
+sys.path.append('batch_shipyard')
+import batch_shipyard.convoy.fleet  # noqa
+print(batch_shipyard.convoy.fleet._ROOT_PATH)
 
 logger = logging.getLogger('trieste')
 
@@ -56,15 +59,15 @@ ClientCollection = namedtuple(
     "ClientCollection",
     'batch_client blob_client queue_client table_client')
 
-
 pass_cli_context = click.make_pass_decorator(TriesteConfig, ensure=True)
 
 
 def _setup_context(config, pool_add_action=False):
-    convoy_fleet.populate_global_settings(config, pool_add_action)
-    clients = convoy_fleet.create_clients(config)
-    convoy_fleet.adjust_general_settings(config)
-    return ClientCollection(clients[0], clients[1], clients[2], clients[3])
+    convoy.fleet.populate_global_settings(config, pool_add_action)
+    clients = convoy.fleet.create_clients(config)
+    convoy.fleet.adjust_general_settings(config)
+    return ClientCollection(*clients)
+
 
 def _config_option(f):
     def callback(ctx, param, value):
@@ -101,48 +104,6 @@ def common_options(f):
     return f
 
 
-def _read_file_to_json(file: str):
-    # type: str -> Maybe[dict]
-    with open(file) as data:
-        try:
-            return Just(json.load(data))
-        except Exception as e:
-            print("Failed to read the file:")
-            return Nothing()
-
-def _to_shipyard_config(
-        config_file,
-        schema_file,
-        converter_func):
-    # type: (str, str, Func) -> Maybe[dict]
-    def convert(x):
-        return Just(converter_func(x))
-
-    return (
-        _read_file_to_json(config_file) >>
-        _validate_json_against_schema(schema_file) >>
-        convert
-    )
-
-@curry
-def _validate_json_against_schema(
-        schema_file: str,
-        _json):
-    # type: str -> Maybe[dict] -> Maybe[dict]
-    with open(schema_file) as schema:
-        _schema = json.load(schema)
-        try:
-            jsonschema.validate(_json, _schema)
-            return Just(_json)
-        except jsonschema.ValidationError as e:
-            print("Failed to validate json with ValidationError:")
-            return Nothing
-        except jsonschema.SchemaError as e:
-            print("Failed to validate json with SchemaError:")
-            return Nothing
-
-
-
 @click.group()
 @click.pass_context
 def cli(ctx):
@@ -173,41 +134,20 @@ def cluster(ctx):
 def cluster_create(ctx, cluster_config, id):
     """Create a cluster with the specified configuration file"""
     inputs = [
-        (ctx.credentials_file,    "api/schema/credentials-schema.json", _to_shipyard_credentials),
-        (ctx.general_config_file, "api/schema/config-schema.json", _to_shipyard_global_config),
-        (cluster_config, "api/schema/cluster-config-schema.json", _to_shipyard_pool_config)
+        (ctx.credentials_file, "api/schema/credentials-schema.json",
+         configurations.to_shipyard_credentials),
+        (ctx.general_config_file, "api/schema/config-schema.json",
+         configurations.to_shipyard_global_config),
+        (cluster_config, "api/schema/cluster-config-schema.json",
+         configurations.to_shipyard_pool_config)
     ]
+    ctx.config = configurations.get_merged_shipyard_config(inputs)
+    ctx.config["pool_specification"]["id"] = id
 
-    def to_config(tuple): 
-        return _to_shipyard_config(tuple[0], tuple[1], tuple[2])
+    print(ctx.config)
+    # clients = _setup_context(ctx.config, True)
+    # print(clients.batch_client)
 
-    configs = map(to_config, inputs)
-
-    for c in configs:
-        print(c)
-
-    # shipyard_credentials = _to_shipyard_config(
-    #     ctx.credentials_file,
-    #     "api/schema/credentials-schema.json",
-    #     _to_shipyard_credentials)
-
-    # shipyard_config = _to_shipyard_config(
-    #     ctx.general_config_file,
-    #     "api/schema/config-schema.json",
-    #     _to_shipyard_global_config)
-    
-    # shipyard_pool_config = _to_shipyard_config(
-    #     cluster_config,
-    #     "api/schema/cluster-config-schema.json",
-    #     _to_shipyard_pool_config)
-
-    # print (shipyard_credentials)
-    # print (shipyard_config)
-    # print (shipyard_pool_config)
-
-    #shipyard_pool_config["pool_specification"]["id"] = id
-
-    # clients = setup_context(ctx.config, True)
     # convoy_fleet.action_pool_add(
     #     clients.batch_client,
     #     clients.blob_client,
@@ -221,117 +161,6 @@ def cluster_create(ctx, cluster_config, id):
 def job(ctx):
     """Job actions"""
     pass
-
-
-# @click.command()
-# @click.option('--config',
-#               default="config.json",
-#               help='(full) path to Configuration JSON file')
-# @click.option('--pool-config',
-#               default="cluster-config.json",
-#               help='(full) path to Pool Configuration JSON file')
-# @click.option('--job-config',
-#               default="job-config.json",
-#               help='(full) path to Job Configuration JSON file')
-# def main(config, pool_config, job_config):
-#     shipyard_config = _to_shipyard_config(
-#         config,
-#         "api/schema/credentials-schema.json",
-#         _to_shipyard_credentials)
-
-#     shipyard_pool_config = _to_shipyard_config(
-#         pool_config,
-#         "api/schema/cluster-config-schema.json",
-#         _to_shipyard_pool_config)
-
-#     shipyard_job_config = _to_shipyard_config(
-#         job_config,
-#         "api/schema/job-config-schema.json",
-#         _to_shipyard_job_config)
-
-#     ctx.config = convoy_util.merge_dict(shipyard_pool_config)
-#     print(shipyard_pool_config.getValue())
-#     print(shipyard_job_config.getValue())
-
-
-
-def _to_shipyard_pool_config(trieste_cluster_config):
-    return {
-        "pool_specification": {
-            "id": trieste_cluster_config["id"],
-            "vm_size": trieste_cluster_config["vm_size"],
-            "vm_count": trieste_cluster_config["vm_count"],
-            "inter_node_communication_enabled": True,
-            "publisher": "Canonical",
-            "offer": "UbuntuServer",
-            "sku": "16.04.0-LTS",
-            "ssh_docker_tunnel": {
-                "username": "docker",
-                "generate_tunnel_script": True
-            },
-            "reboot_on_start_task_failed": True,
-            "block_until_all_global_resources_loaded": True
-        }
-    }
-
-
-def _to_shipyard_job_config(trieste_job_config):
-    config_file_arg = " configFile={}" \
-        .format(trieste_job_config["config_file"])
-    root_dir_arg = " RootDir={}" \
-        .format(trieste_job_config["input_directory"])
-    cntk_cmd = "/cntk/build/cpu/release/bin/cntk{}{}" \
-        .format(config_file_arg, root_dir_arg)
-
-    prep_cmd = "cp -r /cntk/Examples/Image/MNIST/* . "
-    cntk_shell_cmd = '/bin/bash -c "{} && {}"'.format(prep_cmd, cntk_cmd)
-
-    return {
-        "job_specifications": [
-            {
-                "id": trieste_job_config["id"],
-                "tasks": [
-                    {
-                        "image": "alfpark/cntk:cpu-openmpi-mnist-cifar",
-                        "remove_container_after_exit": True,
-                        "command": cntk_shell_cmd
-                    }
-                ]
-            }
-        ]
-    }
-
-
-def _to_shipyard_global_config(trieste_config):
-    return {
-        "batch_shipyard": {
-            "storage_account_settings": "__storage_account_name__",
-            "storage_entity_prefix": "shipyard"
-        },
-        "global_resources": {
-            "docker_images": [
-                "alfpark/cntk:1.7.2-cpu-openmpi-refdata"
-            ]
-        }
-    }
-
-def _to_shipyard_credentials(trieste_credentials_config):
-    return {
-    "credentials": {
-        "batch": {
-            "account": trieste_credentials_config["credentials"]["batch"]["account"],
-            "account_key": trieste_credentials_config["credentials"]["batch"]["account_key"],
-            "account_service_url": trieste_credentials_config["credentials"]["batch"]["account_service_url"]
-        },
-        "storage": {
-            "__storage_account_name__": {
-                "account": trieste_credentials_config["credentials"]["storage"]["account"],
-                "account_key": trieste_credentials_config["credentials"]["storage"]["account_key"],
-                "endpoint": "core.windows.net"
-            }
-        }
-    }
-}
 
 
 if __name__ == '__main__':
